RF: Move environment management functions to tools
